requirements.txt
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
requests>=2.31.0
openai>=1.0.0
httpx>=0.25.0
